+++ src/database/schema.prisma
@@ -13,7 +13,6 @@ model User {
   createdAt DateTime @default(now())
   updatedAt DateTime @updatedAt
   version   Int
-
-  login    String
-  password String
+  login     String
+  password  String
 }
