+++ src/modules/favorite/entities/favorite.entity.ts
@@ -1,0 +1,9 @@
+import { AlbumEntity } from '../../album/entities/album.entity';
+import { ArtistEntity } from '../../artist/entities/artist.entity';
+import { TrackEntity } from '../../track/entities/track.entity';
+
+export interface FavoritesEntity {
+  albums: AlbumEntity[];
+  artists: ArtistEntity[];
+  tracks: TrackEntity[];
+}
