+++ src/modules/favorite/favorite.controller.ts
@@ -9,9 +9,9 @@ import {
   HttpStatus,
 } from '@nestjs/common';
 
-import { AlbumEntity } from '../albums/entities/album.entity';
-import { ArtistEntity } from '../artists/entities/artist.entity';
-import { TrackEntity } from '../tracks/entities/track.entity';
+import { AlbumEntity } from '../album/entities/album.entity';
+import { ArtistEntity } from '../artist/entities/artist.entity';
+import { TrackEntity } from '../track/entities/track.entity';
 import { FavoritesEntity } from './entities/favorite.entity';
 import { FavoriteService } from './favorite.service';
