+++ src/modules/favorite/favorite.module.ts
