+++ src/modules/index.ts
@@ -1,7 +1,7 @@
-export * from './albums/album.module';
-export * from './artists/artist.module';
+export * from './album/album.module';
+export * from './artist/artist.module';
 export * from './database/database.module';
-export * from './favorites/favorite.module';
-export * from './tracks/track.module';
-export * from './users/user.module';
+export * from './favorite/favorite.module';
+export * from './track/track.module';
+export * from './user/user.module';
 export * from './prisma/prisma.module';
